+++ internal/agent/ablation_test.go
@@ -28,13 +28,12 @@ func TestEvidenceAblationStandsDownTheReadinessGate(t *testing.T) {
 }
 
 func TestCompactionAblationCollapsesTheCachePreservingDeferral(t *testing.T) {
-	full := &Agent{contextWindow: 100_000, compactRatio: 0.8}
+	full := &Agent{agentConfig: agentConfig{contextWindow: 100_000, compactRatio: 0.8}}
 	if got := full.compactTrigger(); got != 80_000 {
 		t.Fatalf("control trigger = %d, want 80000", got)
 	}
 
-	off := &Agent{contextWindow: 100_000, compactRatio: 0.8,
-		ablation: ablation.New(ablation.Compaction)}
+	off := &Agent{agentConfig: agentConfig{contextWindow: 100_000, compactRatio: 0.8}, ablation: ablation.New(ablation.Compaction)}
 	// Compaction ablation forces the sole trigger down to 50% so folds fire earlier.
 	if got := off.compactTrigger(); got != 50_000 {
 		t.Fatalf("ablated trigger = %d, want 50000", got)
