+++ internal/agent/path_bound_tools_test.go
@@ -287,11 +287,7 @@ func TestParentWriteReservationBashClaimsWholeWorkspace(t *testing.T) {
 func TestAgentReserveParentWriteSkipsSubagentDepth(t *testing.T) {
 	root := t.TempDir()
 	sched := NewSubagentScheduler(4, 2)
-	a := &Agent{
-		writeScheduler:     sched,
-		writeWorkspaceRoot: root,
-		subagentDepth:      1,
-	}
+	a := &Agent{agentConfig: agentConfig{writeWorkspaceRoot: root, subagentDepth: 1}, writeScheduler: sched}
 	inner := &recordingWriter{name: "write_file"}
 	release, err := a.reserveParentWrite(inner, mustJSON(t, map[string]string{
 		"path": filepath.Join(root, "a.md"), "content": "x",
@@ -309,11 +305,7 @@ func TestAgentReserveParentWriteSkipsSubagentDepth(t *testing.T) {
 func TestAgentReserveParentWriteHoldsClaim(t *testing.T) {
 	root := t.TempDir()
 	sched := NewSubagentScheduler(4, 2)
-	a := &Agent{
-		writeScheduler:     sched,
-		writeWorkspaceRoot: root,
-		subagentDepth:      0,
-	}
+	a := &Agent{agentConfig: agentConfig{writeWorkspaceRoot: root, subagentDepth: 0}, writeScheduler: sched}
 	inner := &recordingWriter{name: "write_file"}
 	release, err := a.reserveParentWrite(inner, mustJSON(t, map[string]string{
 		"path": filepath.Join(root, "a.md"), "content": "x",
